+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.24.2
 
 ### Refactoring
-- Slight refactoring of the design section and variable reorder method
+- Slight refactoring of the design sections and options reorder method
 
 ## 0.24.1 (April 17, 2014)