+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Slight refactoring of the design sections and options reorder method
 - Slight refactoring of `SheetVariable` model
 - Removed unused `remove_file` action from `SheetsController` and `ProjectsController`
+- Removed unused sheet partial
 
 ## 0.24.1 (April 17, 2014)